adjust Netconn.Receive to simplify grpc handler implementation
A new type ADU is defined that contains only the adu slice,
so that only one, not two (overlapping) slices are returned anymore.
This patch also tries to avoid names like "buf" and "msg",
instead starts refering to ADU and PDU (which, btw, is one
byte less than "msg").

Also, Response.Decode is changed to work on the data part only,
i.e. the PDU without the function code. Previously it was defined
to work on the PDU plus the address byte,
which may not be the same as the ADU.

To minimize confusion, response length request options have been
changed to always refer to a PDU (not PDU plus address byte anymore).